+++ ribs/emitters/opt/_evolution_strategy_base.py
@@ -1,0 +1,5 @@
+"""Provides EvolutionStrategyBase."""
+
+
+class EvolutionStrategyBase:
+    pass